[CHG] use uint64 of user-id

+++ service/lessids/lessids.go
@@ -17,7 +17,7 @@ type ResponseJson struct {
 type SessionEntry struct {
 	AccessToken  string    `json:"access_token"`
 	RefreshToken string    `json:"refresh_token"`
-	Uid          uint32    `json:"uid"`
+	Uid          uint64    `json:"uid"`
 	Uuid         string    `json:"uuid"`
 	Uname        string    `json:"uname"`
 	Name         string    `json:"name"`
